# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Djadi Academy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterInput {
  /** @minLength 2 */
  fullName: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type GradeInputGrade = typeof GradeInputGrade[keyof typeof GradeInputGrade];


export const GradeInputGrade = {
  premiere: 'premiere',
  deuxieme: 'deuxieme',
  troisieme: 'troisieme',
} as const;

export interface GradeInput {
  grade: GradeInputGrade;
}

export interface User {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  grade?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt?: string;
}

export interface AuthResponse {
  user: User;
}

export interface Subject {
  id: number;
  name: string;
  nameAr: string;
  nameFr: string;
  grade: string;
  color: string;
  icon: string;
  lessonCount: number;
  /** @nullable */
  description?: string | null;
}

export type LessonType = typeof LessonType[keyof typeof LessonType];


export const LessonType = {
  video: 'video',
  pdf: 'pdf',
  quiz: 'quiz',
} as const;

export interface Lesson {
  id: number;
  title: string;
  titleAr: string;
  subjectId: number;
  /** @nullable */
  subjectName?: string | null;
  grade: string;
  /** Duration in minutes */
  duration: number;
  type: LessonType;
  /** @nullable */
  description?: string | null;
  createdAt?: string;
}

export interface SubjectProgress {
  subjectId: number;
  subjectName: string;
  subjectNameAr: string;
  lessonCount: number;
  color: string;
}

export interface DashboardSummary {
  totalSubjects: number;
  totalLessons: number;
  recentLessons: Lesson[];
  subjectProgress: SubjectProgress[];
}

export type ListLessonsParams = {
subjectId?: number;
};