# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Djadi Academy API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: users
    description: User profile operations
  - name: subjects
    description: Subject and course operations
  - name: lessons
    description: Lesson operations
  - name: dashboard
    description: Dashboard summary operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout the current user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/grade:
    patch:
      operationId: updateGrade
      tags: [users]
      summary: Set or update the user grade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeInput"
      responses:
        "200":
          description: Grade updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List subjects for the current user's grade
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"

  /subjects/{id}:
    get:
      operationId: getSubject
      tags: [subjects]
      summary: Get a subject by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Subject details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /lessons:
    get:
      operationId: listLessons
      tags: [lessons]
      summary: List lessons, optionally filtered by subject
      parameters:
        - name: subjectId
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"

  /lessons/{id}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get a lesson by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Lesson details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats for the current user
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    RegisterInput:
      type: object
      required: [fullName, email, password]
      properties:
        fullName:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 6

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    GradeInput:
      type: object
      required: [grade]
      properties:
        grade:
          type: string
          enum: [premiere, deuxieme, troisieme]

    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      required: [id, fullName, email]
      properties:
        id:
          type: number
        fullName:
          type: string
        email:
          type: string
        grade:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    Subject:
      type: object
      required: [id, name, nameAr, nameFr, grade, color, icon, lessonCount]
      properties:
        id:
          type: number
        name:
          type: string
        nameAr:
          type: string
        nameFr:
          type: string
        grade:
          type: string
        color:
          type: string
        icon:
          type: string
        lessonCount:
          type: number
        description:
          type: ["string", "null"]

    Lesson:
      type: object
      required: [id, title, titleAr, subjectId, grade, duration, type]
      properties:
        id:
          type: number
        title:
          type: string
        titleAr:
          type: string
        subjectId:
          type: number
        subjectName:
          type: ["string", "null"]
        grade:
          type: string
        duration:
          type: number
          description: Duration in minutes
        type:
          type: string
          enum: [video, pdf, quiz]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    DashboardSummary:
      type: object
      required: [totalSubjects, totalLessons, recentLessons, subjectProgress]
      properties:
        totalSubjects:
          type: number
        totalLessons:
          type: number
        recentLessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
        subjectProgress:
          type: array
          items:
            $ref: "#/components/schemas/SubjectProgress"

    SubjectProgress:
      type: object
      required: [subjectId, subjectName, subjectNameAr, lessonCount, color]
      properties:
        subjectId:
          type: number
        subjectName:
          type: string
        subjectNameAr:
          type: string
        lessonCount:
          type: number
        color:
          type: string