# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Djadi Academy API
 * OpenAPI spec version: 0.1.0
 */
import type { Lesson } from './lesson';
import type { SubjectProgress } from './subjectProgress';

export interface DashboardSummary {
  totalSubjects: number;
  totalLessons: number;
  recentLessons: Lesson[];
  subjectProgress: SubjectProgress[];
}