# lib/api-zod/src/generated/types/gradeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Djadi Academy API
 * OpenAPI spec version: 0.1.0
 */
import type { GradeInputGrade } from './gradeInputGrade';

export interface GradeInput {
  grade: GradeInputGrade;
}