# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Djadi Academy API
 * OpenAPI spec version: 0.1.0
 */
import type { LessonType } from './lessonType';

export interface Lesson {
  id: number;
  title: string;
  titleAr: string;
  subjectId: number;
  /** @nullable */
  subjectName?: string | null;
  grade: string;
  /** Duration in minutes */
  duration: number;
  type: LessonType;
  /** @nullable */
  description?: string | null;
  createdAt?: Date;
}