# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Djadi Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /** @minLength 2 */
  fullName: string;
  email: string;
  /** @minLength 6 */
  password: string;
}