# lib/api-zod/src/generated/types/subject.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Djadi Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface Subject {
  id: number;
  name: string;
  nameAr: string;
  nameFr: string;
  grade: string;
  color: string;
  icon: string;
  lessonCount: number;
  /** @nullable */
  description?: string | null;
}