# lib/api-zod/src/generated/types/subjectProgress.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Djadi Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface SubjectProgress {
  subjectId: number;
  subjectName: string;
  subjectNameAr: string;
  lessonCount: number;
  color: string;
}