# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Djadi Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  grade?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt?: Date;
}